MAYBEFIX: locking/mutex: Optimization for donor handoff
Maybe fold this in?

Suleiman noticed when we are proxying a lock owner, and
it releases the lock, we hand the lock over to the donor, but
we don't immediatley preempt into schedule(). Instead the
proxy continues to run for its timeslice using the donor's
context.

Fix this by adding a preempt_schedule() call after we wake
the next lock owner if we were handing it to a donor task.

Need to dig more because the wakeup should have a premept
point, but that can depend on the config

Reported-by: Suleiman Souhlal <suleiman@google.com>
Signed-off-by: John Stultz <jstultz@google.com>